modularize the app
